extract menu in a composable for inheritance replacement
