+++ ui/src/override/components/useLeftMenu.js
@@ -1,0 +1,102 @@
+import {watch, onUpdated, onMounted, ref, computed} from "vue";
+import {useRoute} from "vue-router";
+import {useI18n} from "vue-i18n";
+
+export function useLeftMenu($emit, generatedMenu) {
+    const $route = useRoute()
+    const {locale} = useI18n()
+
+
+
+    function flattenMenu(menu) {
+        return menu.reduce((acc, item) => {
+            if (item.child) {
+                acc.push(...flattenMenu(item.child));
+            }
+
+            acc.push(item);
+            return acc;
+        }, []);
+    }
+
+    function onToggleCollapse(folded) {
+        collapsed.value = folded;
+        localStorage.setItem("menuCollapsed", folded ? "true" : "false");
+        $emit("menu-collapse", folded);
+    }
+
+    function disabledCurrentRoute(items) {
+        return items
+            .map(r => {
+                if (r.href === $route.path) {
+                    r.disabled = true;
+                }
+
+                // route hack is still needed for blueprints
+                if (r.href !== "/" && ($route.path.startsWith(r.href) || r.routes?.includes($route.name))) {
+                    r.class = "vsm--link_active";
+                }
+
+                if (r.child && r.child.some(c => $route.path.startsWith(c.href) || c.routes?.includes($route.name))) {
+                    r.class = "vsm--link_active";
+                    r.child = disabledCurrentRoute(r.child);
+                }
+
+                return r;
+            })
+    }
+
+
+    function expandParentIfNeeded() {
+        document.querySelectorAll(".vsm--link.vsm--link_level-1.vsm--link_active:not(.vsm--link_open)[aria-haspopup]").forEach(e => {
+            e.click()
+        });
+    }
+
+    onUpdated(() => {
+        // Required here because in mounted() the menu is not yet rendered
+        expandParentIfNeeded();
+    })
+
+    watch(locale, () => {
+        localMenu.value = disabledCurrentRoute(generatedMenu.value);
+    }, {deep: true});
+
+    const menu = computed(() => {
+        return disabledCurrentRoute(generatedMenu.value);
+    });
+
+    const $el = ref(null);
+
+    watch(menu, (newVal, oldVal) => {
+              // Check if the active menu item has changed, if yes then update the menu
+              if (JSON.stringify(flattenMenu(newVal).map(e => e.class?.includes("vsm--link_active") ?? false)) !==
+                  JSON.stringify(flattenMenu(oldVal).map(e => e.class?.includes("vsm--link_active") ?? false))) {
+                  localMenu.value = newVal;
+                  $el.value?.querySelectorAll(".vsm--item span").forEach(e => {
+                      //empty icon name on mouseover
+                      e.setAttribute("title", "")
+                  });
+              }
+          },
+          {
+              flush: "post",
+              deep: true
+          });
+
+    const collapsed = ref(localStorage.getItem("menuCollapsed") === "true")
+    const localMenu = ref([])
+
+
+    onMounted(() => {
+        localMenu.value = menu.value;
+    })
+
+    return {
+        collapsed,
+        localMenu,
+        onToggleCollapse,
+        $el,
+        menu
+    }
+}
